Split AppFixtures.php
